+++ src/Messages/Serializer/MessageFieldsEncoder.js
@@ -1,6 +1,6 @@
 import RLP from 'rlp'
-import TEMPLATES from './Serializer/MessageSerializerTemplates.js'
-import PrimitivesEncoder from '../PrimitivesEncoder.js'
+import TEMPLATES from './MessageSerializerTemplates.js'
+import PrimitivesEncoder from '../../PrimitivesEncoder.js'
 
 export default class MessageFieldsEncoder {
     constructor() {
